refactor(core): split types.rs god module into 6 concept submodules (#401)
types.rs had grown to 1946 LOC holding ~46 heterogeneous public types separated
only by `// ---` comments. Split into a `types/` directory by concept:

- types/events.rs     — EventType, Event, NewEvent, EventFilter, Outcome(Counts/Class)
- types/facts.rs      — FactType, Fact, Edge, Summary, New*, AddFact*, ClassifierInput, *Row
- types/scope.rs      — ScopeNode, ScopeQuery
- types/provenance.rs — Promotion/Lineage records, MergeGroup, ConsolidationProposal
- types/cognitive.rs  — Insight, DreamCycleConfig, Promote*, EmbeddingFingerprint, FactId/LineageId
- types/activity.rs   — ActivityStatus, Activity, *Activity*, SessionCheckpoint, ProjectContext

types/mod.rs flat-re-exports each submodule (`pub use events::*; ...`) so the
`crate::types::*` public API is preserved verbatim.

Pure reorganization — no behavior change:
- Public API identical: the 46 exported items are byte-for-byte unchanged
  (verified by a sorted pub-item set diff against the pre-split file).
- All 42 unit tests relocated alongside their types (no test dropped).
- The three private serde-default helpers stay private — `default_root_scope_id`
  moved to facts.rs (its only user); the two event helpers stay module-private in
  events.rs. No new public symbols leak through the glob re-exports.

Gate: build / test (37 binaries, 0 failed) / clippy -D warnings / fmt — all green
under --workspace --all-features.

Fixes #401.
